+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## 3.3.0
   * Dynamically generate graphql query [#214](https://github.com/singer-io/tap-shopify/pull/214)
+  * Introduce new stream order_shipping_lines. [#213](https://github.com/singer-io/tap-shopify/pull/213)
+  * shippingLine is a nested object within the orders stream and it represents individual shipping methods.
 
 ## 3.2.1
   * Refactor refund and transactions sync logic to rely on updated parent objects [#212](https://github.com/singer-io/tap-shopify/pull/212)